DishesController implementation updated: Regular user can manage only his own dishes (additional authorization by claims is used), Moderators and Administrators can manage any data. Dishes not assigned to any user (UserId = null) are available for everyone without authentication and authorization

+++ FinalDish.API/Controllers/DishesController.cs
@@ -1,35 +1,72 @@
 ﻿
 using FinalDish.API.Constants;
 using FinalDish.API.DTO;
+using FinalDish.API.Extensions;
 using FinalDish.API.Models;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 
 namespace FinalDish.API.Controllers
 {
-    [Route("[controller]")]
+    [Route("[controller]/[action]")]
     [ApiController]
     public class DishesController : IntermediateController
     {
         private readonly ApplicationDbContext context;
+        private readonly UserManager<AppUser> userManager;
 
-        public DishesController(ApplicationDbContext context)
+        public DishesController
+            (ApplicationDbContext context, 
+            UserManager<AppUser> userManager)
         {
             this.context = context;
+            this.userManager = userManager;
         }
 
         [HttpGet]
-        [ResponseCache(CacheProfileName = CacheProfilesNames.NoStore)] // TODO: MaxAge300
-        public async Task<IEnumerable<Dish>> Get([FromQuery] RangeRequestDTO data) 
+        [ResponseCache(CacheProfileName = CacheProfilesNames.NoStore)]
+        public async Task<IEnumerable<Dish>> GetBaseDishes([FromQuery] RangeRequestDTO data)
         {
             return await context.Dishes
+                .Where(x => x.UserId == null)
                 .Skip(data.RangeId * data.RangeSize)
                 .Take(data.RangeSize)
                 .ToArrayAsync();
         }
 
         [Authorize(Roles = RolesNames.Moderator)]
+        [HttpGet]
+        [ResponseCache(CacheProfileName = CacheProfilesNames.NoStore)]
+        public async Task<IEnumerable<Dish>> GetAll([FromQuery] RangeRequestDTO data) 
+        {
+            return await context.Dishes
+                .Skip(data.RangeId * data.RangeSize)
+                .Take(data.RangeSize)
+                .ToArrayAsync();
+        }
+        
+        [Authorize]
+        [HttpGet]
+        [ResponseCache(CacheProfileName = CacheProfilesNames.NoStore)]
+        public async Task<IActionResult> GetUserDishes(string? userId)
+        {
+            var user = await userManager.FindByIdAsync(userId);
+
+            if (!IsAuthorizedRequest(user))
+                return Forbid();
+
+            var dishes = await context.Dishes
+                .Where(x => x.UserId == userId)
+                .ToPOD()
+                .ToArrayAsync();
+
+            return JsonResponse(dishes);
+        }
+
+        [Authorize]
         [HttpPost]
         [ResponseCache(CacheProfileName = CacheProfilesNames.NoStore)]
         public async Task<IActionResult> Post(DishDTO data) 
@@ -38,21 +75,25 @@ namespace FinalDish.API.Controllers
             {
                 if (context.DishTypes.Find(data.DishTypeId) is null)
                     return Error500($"Dish type with id = {data.DishTypeId} does not exist");
-                
+
+                var user = await userManager.FindByIdAsync(data.UserId);
+
+                if (data.UserId is not null && user is null)
+                    return Error500($"User with id = {data.UserId} does not exist");
+
+                if (!IsAuthorizedRequest(user))
+                    return Forbid();
+
                 var dish = new Dish 
                 {
                     Name = data.Name,
-                    DishTypeId = data.DishTypeId
+                    DishTypeId = data.DishTypeId,
+                    UserId = data.UserId
                 };
                 var entry = await context.Dishes.AddAsync(dish);
                 await context.SaveChangesAsync();
 
-                return StatusCode(StatusCodes.Status201Created,
-                    new JsonResult(new 
-                    {
-                        DishId = entry.Entity.Id,
-                        Message = "Dish has been created."
-                    }));
+                return Created(string.Empty, $"Dish with id = {entry.Entity.Id} has been created.");
             }
             catch (Exception ex) 
             {
@@ -60,7 +101,7 @@ namespace FinalDish.API.Controllers
             }
         }
 
-        [Authorize(Roles = RolesNames.Moderator)]
+        [Authorize]
         [HttpPut]
         [ResponseCache(CacheProfileName = CacheProfilesNames.NoStore)]
         public async Task<IActionResult> Put(int id, DishDTO data)
@@ -69,6 +110,14 @@ namespace FinalDish.API.Controllers
             {
                 if (context.DishTypes.Find(data.DishTypeId) is null)
                     return Error500($"Dish type with id = {data.DishTypeId} does not exist");
+
+                var user = await userManager.FindByIdAsync(data.UserId);
+
+                if (data.UserId is not null && user is null)
+                    return Error500($"User with id = {data.UserId} does not exist");
+
+                if (!IsAuthorizedRequest(user))
+                    return Forbid();
 
                 var entity = await context.Dishes.FirstOrDefaultAsync(x => x.Id == id);
 
@@ -79,11 +128,11 @@ namespace FinalDish.API.Controllers
 
                     await context.SaveChangesAsync();
 
-                    return StatusCode(StatusCodes.Status200OK, $"Dish with id = {id} has been updated.");
+                    return Ok($"Dish with id = {id} has been updated.");
                 }
                 else 
                 {
-                    return StatusCode(StatusCodes.Status400BadRequest, $"Dish with id = {id} does not exist.");
+                    return BadRequest($"Dish with id = {id} does not exist.");
                 }
             }
             catch (Exception ex)
@@ -96,7 +145,7 @@ namespace FinalDish.API.Controllers
             }
         }
 
-        [Authorize(Roles = RolesNames.Admin)]
+        [Authorize]
         [HttpDelete]
         [ResponseCache(CacheProfileName = CacheProfilesNames.NoStore)]
         public async Task<IActionResult> Delete(int id) 
@@ -107,24 +156,46 @@ namespace FinalDish.API.Controllers
 
                 if (entity is not null)
                 {
+                    var user = await userManager.FindByIdAsync(entity.UserId);
+
+                    if (!IsAuthorizedRequest(user))
+                        return Forbid();
+
                     var entry = context.Dishes.Remove(entity);
                     await context.SaveChangesAsync();
 
-                    return StatusCode(StatusCodes.Status200OK, $"Dish with id = {id} has been removed.");
+                    return Ok($"Dish with id = {id} has been removed.");
                 }
                 else
                 {
-                    return StatusCode(StatusCodes.Status400BadRequest, $"Dish with id = {id} does not exist.");
+                    return BadRequest($"Dish with id = {id} does not exist.");
                 }
             }
             catch (Exception ex) 
             {
-                var details = new ProblemDetails();
-                details.Detail = ex.Message;
-                details.Status = StatusCodes.Status500InternalServerError;
-
-                return StatusCode(StatusCodes.Status500InternalServerError, details);
+                return Error500(ex.Message);
             }
+        }
+
+        private bool IsAuthorizedRequest(AppUser? user)
+        {
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            return IsSupervisorRole(identity) || IsNameMatch(user, identity);
+        }
+
+        private bool IsSupervisorRole(ClaimsIdentity? identity) 
+        {
+            var roles = identity?.FindAll(ClaimTypes.Role)?.Select(x => x.Value);
+
+            return roles is not null && roles.Count() > 0 && roles.Contains(RolesNames.Moderator);
+        }
+
+        private bool IsNameMatch(AppUser? user, ClaimsIdentity? identity) 
+        {
+            var userName = user?.UserName;
+            var nameFromIdentity = identity?.FindFirst(ClaimTypes.Name)?.Value;
+
+            return userName == nameFromIdentity && userName is not null;
         }
     }
 }
